requirements.txt
asgiref==3.6.0
black==23.3.0
click==8.1.3
Django==4.2.1
djangorestframework==3.14.0
mypy-extensions==1.0.0
packaging==23.1
pathspec==0.11.1
Pillow==9.5.0
platformdirs==3.5.1
PyJWT==2.7.0
pytz==2023.3
sqlparse==0.4.4
